#49 refactor(seat): 공연장 전체 좌석 조회 API - DTO 패키지 구조 변경 및 DTO 클래스 record로 변경
