+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/seat/seatstate/LockedSeatListResponse.java
@@ -1,4 +1,4 @@
-package wisoft.nextframe.schedulereservationticketing.dto.seat;
+package wisoft.nextframe.schedulereservationticketing.dto.seat.seatstate;
 
 import java.util.List;
